deep-image-orientation-detection/requirements.txt
torch
torchvision
pillow
numpy
albumentations
tqdm
tensorboard
onnx
onnxruntime